Optionally use pg_roles instead of pg_authid.
pg_roles is a "public" version of pg_authid, which allows use in
more restricted environments (such as RDS).  The downside is
that password comparison doesn't work, as pg_roles only returns
a dummy `********` password, so passwords update logic will be
fired for each account on every run.

Added `--alternate-attributes-table` to activate the use of `pg_role`,
though the default is `--no-alternate-attributes-table` so that
the existing behavior of using `pg_authid` is preserved